chore: split tests for home and mobile phone numbers
