+++ tests/unit/backend/utils/field-validation/home-num-validation.spec.js
@@ -5,17 +5,17 @@ const {
 const {
   ValidateFieldError,
 } = require('../../../../../dist/backend/app/modules/submission/submission.errors')
-describe('Phone number validation tests', () => {
+describe('Home phone number validation tests', () => {
   it('should allow empty answer for required logic field that is not visible', () => {
     const formField = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       required: true,
       allowIntlNumbers: false,
     }
     const response = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       isVisible: false,
       answer: '',
     }
@@ -27,13 +27,13 @@ describe('Phone number validation tests', () => {
   it('should allow empty answer for optional field', () => {
     const formField = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       required: false,
       allowIntlNumbers: false,
     }
     const response = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       isVisible: false,
       answer: '',
     }
@@ -45,13 +45,13 @@ describe('Phone number validation tests', () => {
   it('should not allow empty answer for required field', () => {
     const formField = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       required: true,
       allowIntlNumbers: false,
     }
     const response = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       isVisible: true,
       answer: '',
     }
@@ -60,24 +60,6 @@ describe('Phone number validation tests', () => {
     expect(validateResult._unsafeUnwrapErr()).toEqual(
       new ValidateFieldError('Invalid answer submitted'),
     )
-  })
-
-  it('should allow valid mobile numbers for mobile fieldType', () => {
-    const formField = {
-      _id: 'abc123',
-      fieldType: 'mobile',
-      required: true,
-      allowIntlNumbers: false,
-    }
-    const response = {
-      _id: 'abc123',
-      fieldType: 'mobile',
-      isVisible: false,
-      answer: '+6598765432',
-    }
-    const validateResult = validateField('formId', formField, response)
-    expect(validateResult.isOk()).toBe(true)
-    expect(validateResult._unsafeUnwrap()).toEqual(true)
   })
 
   it('should allow valid home numbers for homeno fieldType', () => {
@@ -97,38 +79,18 @@ describe('Phone number validation tests', () => {
     expect(validateResult._unsafeUnwrap()).toEqual(true)
   })
 
-  it('should disallow mobile numbers without "+" prefix', () => {
+  it('should disallow home numbers without "+" prefix', () => {
     const formField = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       required: true,
       allowIntlNumbers: false,
     }
     const response = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       isVisible: false,
-      answer: '6598765432',
-    }
-    const validateResult = validateField('formId', formField, response)
-    expect(validateResult.isErr()).toBe(true)
-    expect(validateResult._unsafeUnwrapErr()).toEqual(
-      new ValidateFieldError('Invalid answer submitted'),
-    )
-  })
-
-  it('should disallow home numbers on mobile fieldType', () => {
-    const formField = {
-      _id: 'abc123',
-      fieldType: 'mobile',
-      required: true,
-      allowIntlNumbers: false,
-    }
-    const response = {
-      _id: 'abc123',
-      fieldType: 'mobile',
-      isVisible: false,
-      answer: '+6565656565',
+      answer: '6567772918',
     }
     const validateResult = validateField('formId', formField, response)
     expect(validateResult.isErr()).toBe(true)
@@ -160,15 +122,15 @@ describe('Phone number validation tests', () => {
   it('should disallow international numbers when field does not allow for it', () => {
     const formField = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       required: true,
       allowIntlNumbers: false,
     }
     const response = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       isVisible: false,
-      answer: '+447851315617',
+      answer: '+441285291028',
     }
     const validateResult = validateField('formId', formField, response)
     expect(validateResult.isErr()).toBe(true)
@@ -180,15 +142,15 @@ describe('Phone number validation tests', () => {
   it('should allow international numbers when field allows for it', () => {
     const formField = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       required: true,
       allowIntlNumbers: true,
     }
     const response = {
       _id: 'abc123',
-      fieldType: 'mobile',
+      fieldType: 'homeno',
       isVisible: false,
-      answer: '+447851315617',
+      answer: '+441285291028',
     }
     const validateResult = validateField('formId', formField, response)
     expect(validateResult.isOk()).toBe(true)
